Fix compiling zerotier-one binary on Android
Android's version of pthread doesn't have pthread_setattr_default_np,
which this code uses.

According to the man page pthread_getattr_default_np(3), this function
"sets the default attributes [...] used when pthread_create(3) is called
with a second argument that is NULL". A basic search with ripgrep shows
that the only time this occurs is within some code in the attic/
directory. Because of this, I'm not sure this code is even doing
anything on platforms where this function does exist.